Add Password class
Since the password needs to be hashed in several places in the project,
it might be difficult to find all the locations in which the hashing
algorithms are used. Since this is an open source project, people might
want to use their "own" hashing algorithms of choice. This should make
it easier for them to switch the algo and to avoid repetition. It is
also a simple class so it shouldn't conflict our interests.